requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==43.0.0
cssselect2==0.7.0
Django==3.2.10
django-ckeditor==5.9.0
django-ckeditor-filebrowser-filer==0.3.0
django-el-pagination==3.3.0
django-filer==2.2.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-polymorphic==3.1.0
django-registration-redux==2.10
django-versatileimagefield==2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
easy-thumbnails==2.8.5
idna==3.7
lxml==5.2.2
mysqlclient==2.2.4
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-magic==0.4.27
pytz==2024.1
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
Unidecode==1.1.2
urllib3==2.2.2
webencodings==0.5.1